refactor: make checking the difference and monotonicity clearer.
It comes with increased computation cost, because I go through each level 3 times: once to check difference, once to check if it;s increasing and once to check if it's decreasing. But it makes the intention clearer, and the input data is small (1000 lines), so the solution is fast anyway.